+++ HotfixMods.Core/Models/CreatureTemplate.cs
@@ -12,30 +12,30 @@ namespace HotfixMods.Core.Models
         public int MinLevel { get; set; }
         public int MaxLevel { get; set; }
         public int Faction { get; set; }
-        public double Scale { get; set; }
+        public decimal Scale { get; set; }
         [Column("unit_class")]
         public CreatureUnitClasses UnitClass { get; set; }
         public CreatureTypes Type { get; set; }
         public CreatureRanks Rank { get; set; }
         [Column("speed_walk")]
-        public double SpeedWalk { get; set; }
+        public decimal SpeedWalk { get; set; }
         [Column("speed_run")]
-        public double SpeedRun { get; set; }
+        public decimal SpeedRun { get; set; }
         public string? AiName { get; set; }
         public int BaseAttackTime { get; set; }
         public int RangeAttackTime { get; set; }
-        public double BaseVariance { get; set; }
-        public double RangeVariance { get; set; }
-        public double HoverHeight { get; set; }
+        public decimal BaseVariance { get; set; }
+        public decimal RangeVariance { get; set; }
+        public decimal HoverHeight { get; set; }
         public bool RegenHealth { get; set; }
 
-        public double HealthModifier { get; set; }
-        public double HealthModifierExtra { get; set; }
-        public double ManaModifier { get; set; }
-        public double ManaModifierExtra { get; set; }
-        public double ArmorModifier { get; set; }
-        public double DamageModifier { get; set; }
-        public double ExperienceModifier { get; set; }
+        public decimal HealthModifier { get; set; }
+        public decimal HealthModifierExtra { get; set; }
+        public decimal ManaModifier { get; set; }
+        public decimal ManaModifierExtra { get; set; }
+        public decimal ArmorModifier { get; set; }
+        public decimal DamageModifier { get; set; }
+        public decimal ExperienceModifier { get; set; }
 
         [Column("unit_flags")]
         public long UnitFlags { get; set; }
